integrated the user system with the todo routes and controllers

+++ express-todo/controllers/todosController.js
@@ -1,8 +1,9 @@
 const Todo = require("../models/todo");
+const User = require("../models/user");
 const { body, validationResult } = require("express-validator");
 
 exports.todo_all_get = async (req, res) => {
-    const todos = await Todo.find({});
+    const todos = await Todo.find({ user: req.user.id });
 
     if (todos) {
         return res.json({ success: true, result: todos });
@@ -18,6 +19,13 @@ exports.todo_single_get = async (req, res) => {
     const todo = await Todo.findById(req.params.id);
 
     if (todo) {
+        if (req.user.id !== todo.user.toString()) {
+            return res.status(401).json({
+                success: false,
+                message: "You are not authorized to view that",
+            });
+        }
+
         return res.json({ success: true, result: todo });
     } else {
         return res.status(404).json({
@@ -46,14 +54,21 @@ exports.todo_create_post = [
         }
 
         const { title, description, important } = req.body;
+        const user = req.user;
 
         const todo = new Todo({
             title,
             description,
             important,
+            user: user.id,
         });
 
         const newTodo = await todo.save();
+        await User.findByIdAndUpdate(
+            user._id,
+            { $push: { todos: newTodo._id } },
+            { new: true, useFindAndModify: false }
+        );
         return res.status(201).json(newTodo);
     },
 ];
@@ -66,19 +81,30 @@ exports.todo_remove_delete = async (req, res) => {
             success: false,
             message: "Todo not found",
         });
+    }
+
+    if (req.user.id !== todo.user.toString()) {
+        return res.status(401).json({
+            success: false,
+            message: "You are not authorized to delete that",
+        });
+    }
+
+    const deleteTodo = await Todo.findByIdAndDelete(req.params.id);
+    if (deleteTodo) {
+        await User.findByIdAndUpdate(todo.user, {
+            $pull: { todos: req.params.id },
+        });
+
+        return res.json({
+            success: true,
+            message: "Todo successfully deleted",
+        });
     } else {
-        const deleteTodo = await Todo.findByIdAndDelete(req.params.id);
-        if (deleteTodo) {
-            return res.json({
-                success: true,
-                message: "Todo successfully deleted",
-            });
-        } else {
-            return res.status(422).json({
-                success: false,
-                error: "Something went wrong",
-            });
-        }
+        return res.status(422).json({
+            success: false,
+            error: "Something went wrong",
+        });
     }
 };
 
@@ -108,6 +134,13 @@ exports.todo_update_put = [
             return res.status(404).json({
                 success: false,
                 message: "Todo not found",
+            });
+        }
+
+        if (req.user.id !== existingTodo.user.toString()) {
+            return res.status(401).json({
+                success: false,
+                message: "You are not authorized to update that",
             });
         }
